mlp mnist / cosine annealing fix + grad clipping

+++ train.py
@@ -7,6 +7,7 @@ from tqdm import tqdm
 import os
 import argparse
 from cleanfid import fid
+import sys
 
 import matplotlib
 matplotlib.use('TkAgg')
@@ -14,6 +15,8 @@ import matplotlib.pyplot as plt
 
 from models import MLP, load_config, display_results, Path
 from models import UNetModel
+
+sys.setrecursionlimit(1500)
 
 
 def train(
@@ -28,11 +31,18 @@ def train(
         path,
         scheduler_type,
         progress_bar,
+        grad_clip: int = 1.0,
 ):
+    """
+    TIPS: 
+    - decreasing LR may improve the model; oscillation around local minima induces a higher average training error.
+    - if a model seems to be unable to capture the data fully, increase the number of parameters and try a lower LR. 
+    """
 
     optimizer = torch.optim.Adam(model.parameters(), lr=lr)
     if scheduler_type == 'cosine_annealing':
-        scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=10)
+        scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=epochs, eta_min=0.0)
+        # scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=10)
 
     agg_train_loss = []
     agg_val_loss = []
@@ -60,18 +70,25 @@ def train(
             # x_{t} ~ p_{t}(x_{t}; tx_{1}, (1 - (1 - \sigma_{\min})*t)I)
             x_t, u_t = path.sample_path(x_0=x_0, x_1=x_1, t=t)
             
-            # predict the ODE on the transformed sample
             optimizer.zero_grad()
 
             loss = torch.pow(model([x_t, t]) - u_t, 2).mean()
             loss.backward()
+
+            torch.nn.utils.clip_grad_norm_(model.parameters(), grad_clip)
+
             optimizer.step()
 
             total_loss += loss.detach().cpu().item()
 
             # if i % display_train_freq == display_train_freq - 1:
+
         avg_loss = total_loss / len(train_loader)
+        # record average loss per display_train_freq batches
+        # avg_loss = total_loss / display_train_freq
         agg_train_loss.append(avg_loss)
+
+        print(f'\nBatch {i + 1}/{len(train_loader)}\tTrain Loss: {avg_loss:.3e}')
 
         if epoch % display_val_freq == 0:
             with torch.no_grad():
@@ -93,7 +110,6 @@ def train(
             avg_val_loss = total_val_loss / len(val_loader)
             agg_val_loss.append(avg_val_loss)
 
-        print(f'\nBatch: {i + 1} \t Training loss: {avg_loss:.3e}')
         print(f'Validation loss: {avg_val_loss:.3e}')
 
         if scheduler_type is not None:
@@ -122,41 +138,51 @@ def sample(
             t = t.unsqueeze(-1)
         return model([x, t])
 
-    _, axs = plt.subplots(num_samples, n_steps, figsize=(30, 4), sharex=True, sharey=True)
-    axs[0].set_xlim(-3.0, 3.0)
-    axs[0].set_ylim(-3.0, 3.0)
-    
+    if display_plot:
+        _, axs = plt.subplots(num_samples, n_steps, figsize=(30, 4), sharex=True, sharey=True)
+    else:
+        _, axs = plt.subplots(1, num_samples, figsize=(30, 4), sharex=True, sharey=True)
+
     with torch.no_grad():
         for s in range(num_samples):
             # size: B x C x H x W
             x_0 = torch.randn(size=(1, num_channels, *size))
             time_traj = torch.linspace(0, 1, n_steps)
 
-            sol = odeint(func=ode_func, y0=x_0, t=time_traj, atol=1e-4, rtol=1e-4, method=ode_method)
+            sol = odeint(func=ode_func, y0=x_0, t=time_traj, method=ode_method)
 
             if display_plot == 'img':
                 sol = (sol - sol.min()) / (sol.max() - sol.min())
-
                 for i in range(n_steps):
                     # size: n_steps x B x C x H x W, always take n_samples
                     if num_samples != 1:
                         c_ax = axs[s, i]
                     else:
                         c_ax = axs[i]
-                        
+
                     c_ax.imshow(sol[i, 0, :, :, :].permute(1, 2, 0).numpy())
                     c_ax.set_aspect('equal')
                     c_ax.axis('off')
                     c_ax.set_title(f't: {time_traj[i]:.2f}')
             elif display_plot == '2d':
+                # x_{0} ~ p_{x_{0}}(x_{0} | 0, I)
+                axs[0].set_xlim(-3.0, 3.0)
+                axs[0].set_ylim(-3.0, 3.0)
+
                 for i in range(n_steps):
                     c_ax = axs[i]
                     c_ax.scatter(sol[i, :, :, :, 0].numpy(), sol[i, :, :, :, 1].numpy(), s=10)
                     c_ax.set_aspect('equal')
                     c_ax.set_title(f't: {time_traj[i]:.2f}')
+            else:
+                sol = (sol - sol.min()) / (sol.max() - sol.min())
+                axs[s].imshow(sol[-1, 0, :, :, :].permute(1, 2, 0).numpy())
+                axs[s].set_aspect('equal')
+                axs[s].axis('off')
+                axs[s].set_title(f't: {time_traj[-1]:.2f}')
 
-            plt.tight_layout()
-            plt.show()
+        plt.tight_layout()
+        plt.show()
 
     return sol[-1, 0, :, :, :] * 255.
 
@@ -172,27 +198,43 @@ if __name__ == '__main__':
 
     # retrieve config params based on dataset
     dataset_type = config["dataset_type"]
-    data_config = config['data_config'][dataset_type]
+    dataset_config = config['data_config'][dataset_type]
     training = config['training']
+    load_checkpoint = config['load_checkpoint']
 
     # setup config parameters
-    batch_size = data_config['batch_size']
-    num_workers = data_config['num_workers']
-    model_path = data_config['model_path']
-    img_reshape_size = data_config['img_reshape_size']
-    train_val_split = data_config['train_val_split']
-    model_type = data_config['model']
-    epochs = data_config["epochs"]
-    lr = data_config["learning_rate"]
-    display_train_freq = data_config["display_train_freq"]
-    display_val_freq = data_config["display_val_freq"]
-    num_samples = data_config['num_samples']
-    n_steps = data_config['inference_steps']
-    sigma_min = data_config['sigma_min']
-    ode_method = data_config['ode_method']
-    scheduler_type = data_config['scheduler_type']
-    path_type = data_config['path_type']
-    progress_bar = data_config['progress_bar']
+    batch_size = dataset_config['batch_size']
+    num_workers = dataset_config['num_workers']
+    model_path = dataset_config['model_path']
+    img_reshape_size = dataset_config['img_reshape_size']
+    train_val_split = dataset_config['train_val_split']
+    model_type = dataset_config['model']
+    epochs = dataset_config["epochs"]
+    lr = dataset_config["learning_rate"]
+    display_train_freq = dataset_config["display_train_freq"]
+    display_val_freq = dataset_config["display_val_freq"]
+    num_samples = dataset_config['num_samples']
+    n_steps = dataset_config['inference_steps']
+    sigma_min = dataset_config['sigma_min']
+    ode_method = dataset_config['ode_method']
+    scheduler_type = dataset_config['scheduler_type']
+    path_type = dataset_config['path_type']
+    progress_bar = dataset_config['progress_bar']
+    display_plot = dataset_config['display_plot']
+    
+    # model-related params
+    model_config = dataset_config["model_config"]
+    mlp_hidden_dim = model_config["mlp_hidden_dim"]
+    enable_time_embed = model_config["enable_time_embed"]
+    num_hidden_layers = model_config["num_hidden_layers"]
+    out_time_dim = model_config["out_time_dim"]
+    num_res_blocks = model_config["num_res_blocks"]
+    num_channels = model_config["num_channels"]
+    channel_mult = model_config["channel_mult"]
+    num_heads = model_config["num_heads"]
+    num_head_channels = model_config["num_head_channels"]
+    attention_resolutions = model_config["attention_resolutions"]
+    dropout = model_config["dropout"]
 
     # normalize on image channels depending on dataset
     transform = transforms.Compose([
@@ -215,7 +257,6 @@ if __name__ == '__main__':
         rgb_channels = len(dataset[0][0])
         in_dim = np.prod(img_reshape_size) * rgb_channels
         size = img_reshape_size
-        display_plot = 'img'
 
     elif dataset_type == 'mnist':
         transform.transforms.append(transforms.Normalize(mean=0.5, std=0.5))
@@ -237,7 +278,6 @@ if __name__ == '__main__':
         rgb_channels = len(dataset[0][0])
         in_dim = np.prod(img_reshape_size) * rgb_channels
         size = img_reshape_size
-        display_plot = 'img'
 
     elif dataset_type == 'moons':
         from sklearn.datasets import make_moons
@@ -269,40 +309,36 @@ if __name__ == '__main__':
         # sampling
         size = (num_sample_points, in_dim)
         rgb_channels = 1
-        display_plot = '2d'
 
     if model_type == 'mlp':
         model = MLP(
             in_dim=in_dim,
             num_channels=rgb_channels,
+            h_dim=mlp_hidden_dim,
+            enable_time_embed=enable_time_embed,
+            num_hidden=num_hidden_layers,
+            out_time_dim=out_time_dim,
         ).to(device)
     elif model_type == 'unet':
-        # TODO: setup model config parameters
-        if dataset_type == 'mnist':
-            model = UNetModel(
-                dim=(rgb_channels, *img_reshape_size), 
-                num_channels=32, 
-                num_res_blocks=1,
-                # num_classes=10,
-                class_cond=False,
-            ).to(device)
-        elif dataset_type == 'cifar10':
-            model = UNetModel(
-                dim=(rgb_channels, *img_reshape_size),
-                num_res_blocks=2,
-                num_channels=32,
-                channel_mult=[1, 2, 2, 2],
-                num_heads=4,
-                num_head_channels=64,
-                attention_resolutions="16",
-                dropout=0.1,
-                class_cond=False,
-            ).to(device)
+        model = UNetModel(
+            dim=(rgb_channels, *img_reshape_size),
+            num_res_blocks=num_res_blocks,
+            num_channels=num_channels,
+            channel_mult=channel_mult,
+            num_heads=num_heads,
+            num_head_channels=num_head_channels,
+            attention_resolutions=attention_resolutions,
+            dropout=dropout,
+            class_cond=False,
+        ).to(device)
 
     if training:
         # reset model checkpoint
-        if os.path.isfile(model_path):
-            os.remove(model_path)
+        if load_checkpoint:
+            model.load_state_dict(torch.load(model_path, weights_only=True))
+        else:
+            if os.path.isfile(model_path):
+                os.remove(model_path)
 
         from torch.utils.data import DataLoader
